aevorex-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx>=0.28.1
loguru==0.7.2
openai>=1.40.0
google-generativeai>=0.8.0
google-genai>=1.36.0
requests>=2.31.0
Pillow>=11.0.0
